Small Fixes and Emoji Addition
Fixed gallery button inconsistency between safari and firefox.
Fixed misaligned text in the contacts page.
Cleared out the rest of light-dark functions out of the database excluding for the utilities file.
Fixed left nav drawer to but height accurate with a parametric feature to adjust it.
Added emojis beside headers and a toggle to turn it on and off.